update whole structure
