+++ brainunit/math/_fun_accept_unitless.py
@@ -24,7 +24,7 @@ from .._misc import set_module_as
 
 __all__ = [
   # math funcs only accept unitless (unary)
-  'exp', 'exp2', 'expm1', 'log', 'log10', 'log1p', 'log2',
+  'exprel', 'exp', 'exp2', 'expm1', 'log', 'log10', 'log1p', 'log2',
   'arccos', 'arccosh', 'arcsin', 'arcsinh', 'arctan',
   'arctanh', 'cos', 'cosh', 'sin', 'sinc', 'sinh', 'tan',
   'tanh', 'deg2rad', 'rad2deg', 'degrees', 'radians', 'angle',
@@ -72,6 +72,79 @@ def _fun_accept_unitless_unary(
   else:
     assert unit_to_scale is None, f'Unit should be None for the function "{func}" when "x" is not a Quantity.'
     return func(x, *args, **kwargs)
+
+
+def _exprel_v1(x):  # This approximation has problems of the gradient vanishing at x=0
+  # following the implementation of exprel from scipy.special
+  x = jnp.asarray(x)
+  dtype = x.dtype
+
+  # Adjust the tolerance based on the dtype of x
+  if dtype == jnp.float64:
+    small_threshold = 1e-16
+    big_threshold = 717
+  elif dtype == jnp.float32:
+    small_threshold = 1e-8
+    big_threshold = 100
+  elif dtype == jnp.float16:
+    small_threshold = 1e-4
+    big_threshold = 10
+  else:
+    small_threshold = 1e-4
+    big_threshold = 10
+
+  small = jnp.abs(x) < small_threshold
+  big = x > big_threshold
+  origin = jnp.expm1(x) / x
+  return jnp.where(small, 1.0, jnp.where(big, jnp.inf, origin))
+
+
+def _exprel_v2(x, *, level: int = 2):
+  x = jnp.asarray(x)
+  dtype = x.dtype
+  assert jnp.issubdtype(dtype, jnp.floating), f'The input array must contain real numbers. Got {x}'
+
+  # Adjust the tolerance based on the dtype of x
+  if dtype == jnp.float64:
+    threshold = 1e-8
+  elif dtype == jnp.float32:
+    threshold = 1e-5
+  elif dtype in [jnp.float16, jnp.bfloat16]:
+    threshold = 1e-3
+  else:
+    threshold = 1e-3
+
+  assert level in [0, 1, 2, 3], 'The approximation level should be 0, 1, 2, or 3.'
+  if level == 0:
+    return jax.numpy.where(jnp.abs(x) <= threshold, 1., jnp.expm1(x) / x)
+  elif level == 1:
+    return jax.numpy.where(jnp.abs(x) <= threshold, 1. + x / 2., jnp.expm1(x) / x)
+  elif level == 2:
+    return jax.numpy.where(jnp.abs(x) <= threshold, 1. + x / 2. + x * x / 6., jnp.expm1(x) / x)
+  elif level == 3:
+    x2 = x * x
+    return jax.numpy.where(jnp.abs(x) <= threshold, 1. + x / 2. + x2 / 6. + x2 * x / 24., jnp.expm1(x) / x)
+  else:
+    raise ValueError(f'Unsupported approximation level {level}.')
+
+
+@set_module_as('brainunit.math')
+def exprel(x, *, level: int = 2):
+  """
+  Relative error exponential, ``(exp(x) - 1)/x``.
+
+  When ``x`` is near zero, ``exp(x)`` is near 1, so the numerical calculation of ``exp(x) - 1`` can
+  suffer from catastrophic loss of precision. ``exprel(x)`` is implemented to avoid the loss of
+  precision that occurs when ``x`` is near zero.
+
+  Args:
+    x: ndarray. Input array. ``x`` must contain real numbers.
+    level: int. The approximation level of the function. The higher the level, the more accurate the result.
+
+  Returns:
+    ``(exp(x) - 1)/x``, computed element-wise.
+  """
+  return _fun_accept_unitless_unary(_exprel_v2, x, level=level)
 
 
 @set_module_as('brainunit.math')
