--- brainunit/math/_numpy_array_manipulation.py
@@ -1,1505 +1,0 @@
-# Copyright 2024 BDP Ecosystem Limited. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-from __future__ import annotations
-
-import functools
-from collections.abc import Sequence
-from typing import (Union, Optional, Tuple, List)
-
-import jax
-import jax.numpy as jnp
-from jax import Array
-
-from ._numpy_keep_unit import _fun_keep_unit_unary
-from ._numpy_remove_unit import _fun_remove_unit_unary
-from .._base import Quantity, fail_for_dimension_mismatch
-from .._misc import set_module_as
-
-__all__ = [
-  # array manipulation
-  'reshape', 'moveaxis', 'transpose', 'swapaxes', 'row_stack',
-  'concatenate', 'stack', 'vstack', 'hstack', 'dstack', 'column_stack',
-  'split', 'array_split', 'dsplit', 'hsplit', 'vsplit', 'tile', 'repeat', 'unique',
-  'append', 'flip', 'fliplr', 'flipud', 'roll', 'atleast_1d', 'atleast_2d',
-  'atleast_3d', 'expand_dims', 'squeeze', 'sort', 'argsort', 'argmax', 'argmin',
-  'nanargmax', 'nanargmin',
-  'argwhere', 'nonzero', 'flatnonzero', 'searchsorted', 'extract',
-  'count_nonzero', 'max', 'min', 'amax', 'amin', 'block', 'compress',
-  'diagflat', 'diagonal', 'choose', 'ravel',
-]
-
-
-# array manipulation
-# ------------------
-
-
-@set_module_as('brainunit.math')
-def reshape(
-    a: Union[Array, Quantity],
-    shape: Union[int, Tuple[int, ...]],
-    order: str = 'C'
-) -> Union[Array, Quantity]:
-  """
-  Gives a new shape to a quantity or an array without changing its data.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Array to be reshaped.
-  shape : int or tuple of ints
-    The new shape should be compatible with the original shape. If
-    an integer, then the result will be a 1-D array of that length.
-    One shape dimension can be -1. In this case, the value is
-    inferred from the length of the array and remaining dimensions.
-  order : {'C', 'F', 'A'}, optional
-    Read the elements of `a` using this index order, and place the
-    elements into the reshaped array using this index order.  'C'
-    means to read / write the elements using C-like index order,
-    with the last axis index changing fastest, back to the first
-    axis index changing slowest. 'F' means to read / write the
-    elements using Fortran-like index order, with the first index
-    changing fastest, and the last index changing slowest. Note that
-    the 'C' and 'F' options take no account of the memory layout of
-    the underlying array, and only refer to the order of indexing.
-    'A' means to read / write the elements in Fortran-like index
-    order if `a` is Fortran *contiguous* in memory, C-like order
-    otherwise.
-
-  Returns
-  -------
-  reshaped_array : ndarray, Quantity
-    This will be a new view object if possible; otherwise, it will
-    be a copy.  Note there is no guarantee of the *memory layout* (C- or
-    Fortran- contiguous) of the returned array.
-  """
-  return _fun_keep_unit_unary(jnp.reshape, a, shape=shape, order=order)
-
-
-@set_module_as('brainunit.math')
-def moveaxis(
-    a: Union[Array, Quantity],
-    source: Union[int, Tuple[int, ...]],
-    destination: Union[int, Tuple[int, ...]]
-) -> Union[Array, Quantity]:
-  """
-  Moves axes of a quantity or an array to new positions.
-  Other axes remain in their original order.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    The array whose axes should be reordered.
-  source : int or sequence of int
-    Original positions of the axes to move. These must be unique.
-  destination : int or sequence of int
-    Destination positions for each of the original axes. These must also be
-    unique.
-
-  Returns
-  -------
-  result : ndarray, Quantity
-    Array with moved axes. This array is a view of the input array.
-  """
-  return _fun_keep_unit_unary(jnp.moveaxis, a, source=source, destination=destination)
-
-
-@set_module_as('brainunit.math')
-def transpose(
-    a: Union[Array, Quantity],
-    axes: Optional[Union[int, Tuple[int, ...]]] = None
-) -> Union[Array, Quantity]:
-  """
-  Permute the dimensions of a quantity or an array.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Input array.
-  axes : list of ints, optional
-    By default, reverse the dimensions, otherwise permute the axes
-    according to the values given.
-
-  Returns
-  -------
-  p : ndarray, Quantity
-    `a` with its axes permuted.  A view is returned whenever
-    possible.
-  """
-  return _fun_keep_unit_unary(jnp.transpose, a, axes=axes)
-
-
-@set_module_as('brainunit.math')
-def swapaxes(
-    a: Union[Array, Quantity],
-    axis1: int,
-    axis2: int
-) -> Union[Array, Quantity]:
-  """
-  Interchange two axes of a quantity or an array.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Input array.
-  axis1 : int
-    First axis.
-  axis2 : int
-    Second axis.
-
-  Returns
-  -------
-  a_swapped : ndarray, Quantity
-    a new array where the axes are swapped.
-  """
-  return _fun_keep_unit_unary(jnp.swapaxes, a, axis1=axis1, axis2=axis2)
-
-
-@set_module_as('brainunit.math')
-def tile(
-    A: Union[Array, Quantity],
-    reps: Union[int, Tuple[int, ...]]
-) -> Union[Array, Quantity]:
-  """
-  Construct a quantity or an array by repeating A the number of times given by reps.
-
-  Parameters
-  ----------
-  A : array_like, Quantity
-    The input array.
-  reps : array_like
-    The number of repetitions of A along each axis.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    The tiled output array.
-  """
-  return _fun_keep_unit_unary(jnp.tile, A, reps=reps)
-
-
-@set_module_as('brainunit.math')
-def repeat(
-    a: Union[Array, Quantity],
-    repeats: Union[int, Tuple[int, ...]],
-    axis: Optional[int] = None,
-    total_repeat_length: Optional[int] = None
-) -> Union[Array, Quantity]:
-  """
-  Repeat elements of a quantity or an array.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Input array.
-  repeats : int or tuple of ints
-    The number of repetitions for each element. `repeats` is broadcasted to fit the shape of the given axis.
-  axis : int, optional
-    The axis along which to repeat values. By default, use the flattened input array, and return a flat output array.
-  total_repeat_length : int, optional
-    The total length of the repeated array. If `total_repeat_length` is not None, the output array
-    will have the length of `total_repeat_length`.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    Output array which has the same shape as `a`, except along the given axis.
-  """
-  return _fun_keep_unit_unary(jnp.repeat, a, repeats=repeats, axis=axis, total_repeat_length=total_repeat_length)
-
-
-@set_module_as('brainunit.math')
-def flip(
-    m: Union[Array, Quantity],
-    axis: Optional[Union[int, Tuple[int, ...]]] = None
-) -> Union[Array, Quantity]:
-  """
-  Reverse the order of elements in a quantity or an array along the given axis.
-
-  Parameters
-  ----------
-  m : array_like, Quantity
-    Input array.
-  axis : int or tuple of ints, optional
-    Axis or axes along which to flip over. The default, axis=None, will flip over all of the axes of the input array.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    A view of `m` with the entries of axis reversed. Since a view is returned, this operation is done in constant time.
-  """
-  return _fun_keep_unit_unary(jnp.flip, m, axis=axis)
-
-
-@set_module_as('brainunit.math')
-def fliplr(
-    m: Union[Array, Quantity]
-) -> Union[Array, Quantity]:
-  """
-  Flip quantity or array in the left/right direction.
-
-  Parameters
-  ----------
-  m : array_like, Quantity
-    Input array.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    A view of `m` with the columns reversed. Since a view is returned, this operation is done in constant time.
-  """
-  return _fun_keep_unit_unary(jnp.fliplr, m)
-
-
-@set_module_as('brainunit.math')
-def flipud(
-    m: Union[Array, Quantity]
-) -> Union[Array, Quantity]:
-  """
-  Flip quantity or array in the up/down direction.
-
-  Parameters
-  ----------
-  m : array_like, Quantity
-    Input array.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    A view of `m` with the rows reversed.
-  """
-  return _fun_keep_unit_unary(jnp.flipud, m)
-
-
-@set_module_as('brainunit.math')
-def roll(
-    a: Union[Array, Quantity],
-    shift: Union[int, Tuple[int, ...]],
-    axis: Optional[Union[int, Tuple[int, ...]]] = None
-) -> Union[Array, Quantity]:
-  """
-  Roll quantity or array elements along a given axis.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Input array.
-  shift : int or tuple of ints
-    The number of places by which elements are shifted. If a tuple, then `axis` must be a tuple of the same size,
-    and each of the given axes is shifted by the corresponding number. If an int while `axis` is a tuple of ints,
-    then the same value is used for all given axes.
-  axis : int or tuple of ints, optional
-    Axis or axes along which elements are shifted. By default, the array is flattened before shifting, after which
-    the original shape is restored.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    Output array, with the same shape as `a`.
-  """
-  return _fun_keep_unit_unary(jnp.roll, a, shift=shift, axis=axis)
-
-
-@set_module_as('brainunit.math')
-def expand_dims(
-    a: Union[Array, Quantity],
-    axis: int
-) -> Union[Array, Quantity]:
-  """
-  Expand the shape of a quantity or an array.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Input array.
-  axis : int
-    Position in the expanded axes where the new axis is placed.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    View of `a` with the number of dimensions increased by one.
-  """
-  return _fun_keep_unit_unary(jnp.expand_dims, a, axis=axis)
-
-
-@set_module_as('brainunit.math')
-def squeeze(
-    a: Union[Array, Quantity],
-    axis: Optional[Union[int, Tuple[int, ...]]] = None
-) -> Union[Array, Quantity]:
-  """
-  Remove single-dimensional entries from the shape of a quantity or an array.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Input data.
-  axis : None or int or tuple of ints, optional
-    Selects a subset of the single-dimensional entries in the shape. If an axis is selected with shape entry greater
-    than one, an error is raised.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    An array with the same data as `a`, but with a lower dimension.
-  """
-  return _fun_keep_unit_unary(jnp.squeeze, a, axis=axis)
-
-
-@set_module_as('brainunit.math')
-def sort(
-    a: Union[Array, Quantity],
-    axis: Optional[int] = -1,
-    *,
-    kind: None = None,
-    order: None = None,
-    stable: bool = True,
-    descending: bool = False,
-) -> Union[Array, Quantity]:
-  """
-  Return a sorted copy of a quantity or an array.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Array or quantity to be sorted.
-  axis : int or None, optional
-    Axis along which to sort. If None, the array is flattened before sorting. The default is -1, which sorts along
-    the last axis.
-  kind : {'quicksort', 'mergesort', 'heapsort', 'stable'}, optional
-    Sorting algorithm. The default is 'quicksort'.
-  order : str or list of str, optional
-    When `a` is a quantity, it can be a string or a sequence of strings, which is interpreted as an order the quantity
-    should be sorted. The default is None.
-  stable : bool, optional
-    Whether to use a stable sorting algorithm. The default is True.
-  descending : bool, optional
-    Whether to sort in descending order. The default is False.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    Sorted copy of the input array.
-  """
-  return _fun_keep_unit_unary(jnp.sort, a, axis=axis, kind=kind, order=order, stable=stable, descending=descending)
-
-
-@set_module_as('brainunit.math')
-def max(
-    a: Union[Array, Quantity],
-    axis: Optional[int] = None,
-    keepdims: bool = False,
-    initial: Optional[Union[int, float, Quantity]] = None,
-    where: Optional[Array] = None,
-) -> Union[Array, Quantity]:
-  """
-  Return the maximum of a quantity or an array or maximum along an axis.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Array or quantity containing numbers whose maximum is desired.
-  axis : int or None, optional
-    Axis or axes along which to operate. By default, flattened input is used.
-  keepdims : bool, optional
-    If this is set to True, the axes which are reduced are left in the result as dimensions with size one. With this
-    option, the result will broadcast correctly against the input array.
-  initial : scalar, optional
-    The minimum value of an output element. Must be present to allow computation on empty slice.
-    See `numpy.ufunc.reduce`.
-  where : array_like, optional
-    Values of True indicate to calculate the ufunc at that position, values of False indicate to leave the value in the
-    output alone.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    Maximum of `a`. If `axis` is None, the result is a scalar value. If `axis` is given, the result is an array of
-    dimension `a.ndim - 1`.
-  """
-  return _fun_keep_unit_unary(jnp.max, a, axis=axis, keepdims=keepdims, initial=initial, where=where)
-
-
-@set_module_as('brainunit.math')
-def min(
-    a: Union[Array, Quantity],
-    axis: Optional[int] = None,
-    keepdims: bool = False,
-    initial: Optional[Union[int, float, Quantity]] = None,
-    where: Optional[Array] = None,
-) -> Union[Array, Quantity]:
-  """
-  Return the minimum of a quantity or an array or minimum along an axis.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Array or quantity containing numbers whose minimum is desired.
-  axis : int or None, optional
-    Axis or axes along which to operate. By default, flattened input is used.
-  keepdims : bool, optional
-    If this is set to True, the axes which are reduced are left in the result as dimensions with size one. With this
-    option, the result will broadcast correctly against the input array.
-  initial : scalar, optional
-    The maximum value of an output element. Must be present to allow computation on empty slice.
-    See `numpy.ufunc.reduce`.
-  where : array_like, optional
-    Values of True indicate to calculate the ufunc at that position, values of False indicate to leave the value in the
-    output alone.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    Minimum of `a`. If `axis` is None, the result is a scalar value. If `axis` is given, the result is an array of
-    dimension `a.ndim - 1`.
-  """
-  return _fun_keep_unit_unary(jnp.min, a, axis=axis, keepdims=keepdims, initial=initial, where=where)
-
-
-amax = max
-amin = min
-
-
-@set_module_as('brainunit.math')
-def diagonal(
-    a: Union[jax.Array, Quantity],
-    offset: int = 0,
-    axis1: int = 0,
-    axis2: int = 1
-) -> Union[jax.Array, Quantity]:
-  """
-  Return specified diagonals.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Array from which the diagonals are taken.
-  offset : int, optional
-    Offset of the diagonal from the main diagonal. Can be positive or negative. Defaults to main diagonal (0).
-  axis1 : int, optional
-    Axis to be used as the first axis of the 2-D sub-arrays from which the diagonals should be taken. Defaults to first
-    axis (0).
-  axis2 : int, optional
-    Axis to be used as the second axis of the 2-D sub-arrays from which the diagonals should be taken. Defaults to
-    second axis (1).
-
-  Returns
-  -------
-  res : ndarray
-    The extracted diagonals. The shape of the output is determined by considering the shape of the input array with
-    the specified axis removed.
-  """
-  return _fun_keep_unit_unary(jnp.diagonal, a, offset=offset, axis1=axis1, axis2=axis2)
-
-
-@set_module_as('brainunit.math')
-def ravel(
-    a: Union[jax.Array, Quantity],
-    order: str = 'C'
-) -> Union[jax.Array, Quantity]:
-  """
-  Return a contiguous flattened quantity or array.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Input array. The elements in `a` are read in the order specified by `order`, and packed as a 1-D array.
-  order : {'C', 'F', 'A', 'K'}, optional
-    The elements of `a` are read using this index order. 'C' means to index the elements in row-major, C-style order,
-    with the last axis index changing fastest, back to the first axis index changing slowest. 'F' means to index the
-    elements in column-major, Fortran-style order, with the first index changing fastest, and the last index changing
-    slowest. 'A' means to read the elements in Fortran-like index order if `a` is Fortran contiguous in memory, C-like
-    order otherwise. 'K' means to read the elements in the order they occur in memory, except for reversing the data
-    when strides are negative. By default, 'C' index order is used.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    The flattened quantity or array. The shape of the output is the same as `a`, but the array is 1-D.
-  """
-  return _fun_keep_unit_unary(jnp.ravel, a, order=order)
-
-
-@set_module_as('brainunit.math')
-def choose(
-    a: Union[Array, Quantity],
-    choices: Sequence[Union[Array, Quantity]],
-    mode: str = 'raise',
-) -> Union[Array, Quantity]:
-  """
-  Construct a quantity or an array from an index array and a set of arrays to choose from.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    This array must be an integer array of the same shape as `choices`. The elements of `a` are used to select elements
-    from `choices`.
-  choices : sequence of array_like, Quantity
-    Choice arrays. `a` and all `choices` must be broadcastable to the same shape.
-  mode : {'raise', 'wrap', 'clip'}, optional
-    Specifies how indices outside [0, n-1] will be treated:
-    - 'raise' : raise an error (default)
-    - 'wrap' : wrap around
-    - 'clip' : clip to the range [0, n-1]
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    The constructed array. The shape is identical to the shape of `a`, and the data type is the data type of `choices`.
-  """
-  return _fun_keep_unit_unary(jnp.choose, a, choices=choices, mode=mode)
-
-
-@set_module_as('brainunit.math')
-def diagflat(
-    v: Union[Array, Quantity],
-    k: int = 0
-) -> Union[Array, Quantity]:
-  """
-  Create a two-dimensional a quantity or array with the flattened input as a diagonal.
-
-  Parameters
-  ----------
-  v : array_like, Quantity
-    Input data, which is flattened and set as the `k`-th diagonal of the output.
-  k : int, optional
-    Diagonal in question. The default is 0.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    The 2-D output array.
-  """
-  return _fun_keep_unit_unary(jnp.diagflat, v, k=k)
-
-
-def _fun_keep_unit_sequence(
-    func,
-    *args,
-    **kwargs
-):
-  leaves, treedef = jax.tree.flatten(args, is_leaf=lambda x: isinstance(x, Quantity))
-  dims = [x.dim if isinstance(x, Quantity) else None for x in leaves]
-  if not all(dim == dims[0] for dim in dims[1:]):
-    raise ValueError(f'Units in args do not match for {func.__name__} operation. Got {dims}.')
-  leaves = [x.value if isinstance(x, Quantity) else x for x in leaves]
-  args = treedef.unflatten(leaves)
-  r = func(*args, **kwargs)
-  if dims[0] is not None:
-    return Quantity(r, dim=dims[0])
-  return r
-
-
-@set_module_as('brainunit.math')
-def concatenate(
-    arrays: Union[Sequence[Array], Sequence[Quantity]],
-    axis: Optional[int] = None,
-    dtype: Optional[jax.typing.DTypeLike] = None
-) -> Union[Array, Quantity]:
-  """
-  Join a sequence of quantities or arrays along an existing axis.
-
-  Parameters
-  ----------
-  arrays : sequence of array_like, Quantity
-    The arrays must have the same shape, except in the dimension corresponding
-    to `axis` (the first, by default).
-  axis : int, optional
-    The axis along which the arrays will be joined.  Default is 0.
-  dtype : dtype, optional
-    If provided, the concatenation will be done using this dtype. Otherwise, the
-    array with the highest precision will be used.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    The concatenated array. The type of the array is the same as that of the
-    first array passed in.
-  """
-  return _fun_keep_unit_sequence(jnp.concatenate, arrays, axis=axis, dtype=dtype)
-
-
-@set_module_as('brainunit.math')
-def stack(
-    arrays: Union[Sequence[Array], Sequence[Quantity]],
-    axis: int = 0,
-    dtype: Optional[jax.typing.DTypeLike] = None
-) -> Union[Array, Quantity]:
-  """
-  Join a sequence of quantities or arrays along a new axis.
-
-  Parameters
-  ----------
-  arrays : sequence of array_like, Quantity
-    The arrays must have the same shape.
-  axis : int, optional
-    The axis in the result array along which the input arrays are stacked.
-  dtype : dtype, optional
-    If provided, the concatenation will be done using this dtype. Otherwise, the
-    array with the highest precision will be used.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    The stacked array has one more dimension than the input arrays.
-  """
-  return _fun_keep_unit_sequence(jnp.stack, arrays, axis=axis, dtype=dtype)
-
-
-@set_module_as('brainunit.math')
-def vstack(
-    tup: Union[Sequence[Array], Sequence[Quantity]],
-    dtype: Optional[jax.typing.DTypeLike] = None
-) -> Union[Array, Quantity]:
-  """
-  Stack quantities or arrays in sequence vertically (row wise).
-
-  Parameters
-  ----------
-  tup : sequence of array_like, Quantity
-    The arrays must have the same shape along all but the first axis.
-  dtype : dtype, optional
-    If provided, the concatenation will be done using this dtype. Otherwise, the
-    array with the highest precision will be used.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    The array formed by stacking the given arrays.
-  """
-  return _fun_keep_unit_sequence(jnp.vstack, tup, dtype=dtype)
-
-
-row_stack = vstack
-
-
-@set_module_as('brainunit.math')
-def hstack(
-    arrays: Union[Sequence[Array], Sequence[Quantity]],
-    dtype: Optional[jax.typing.DTypeLike] = None
-) -> Union[Array, Quantity]:
-  """
-  Stack quantities arrays in sequence horizontally (column wise).
-
-  Parameters
-  ----------
-  arrays : sequence of array_like, Quantity
-    The arrays must have the same shape along all but the second axis.
-  dtype : dtype, optional
-    If provided, the concatenation will be done using this dtype. Otherwise, the
-    array with the highest precision will be used.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    The array formed by stacking the given arrays.
-  """
-  return _fun_keep_unit_sequence(jnp.hstack, arrays, dtype=dtype)
-
-
-@set_module_as('brainunit.math')
-def dstack(
-    arrays: Union[Sequence[Array], Sequence[Quantity]],
-    dtype: Optional[jax.typing.DTypeLike] = None
-) -> Union[Array, Quantity]:
-  """
-  Stack quantities or arrays in sequence depth wise (along third axis).
-
-  Parameters
-  ----------
-  arrays : sequence of array_like, Quantity
-    The arrays must have the same shape along all but the third axis.
-  dtype : dtype, optional
-    If provided, the concatenation will be done using this dtype. Otherwise, the
-    array with the highest precision will be used.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    The array formed by stacking the given arrays.
-  """
-  return _fun_keep_unit_sequence(jnp.dstack, arrays, dtype=dtype)
-
-
-@set_module_as('brainunit.math')
-def column_stack(
-    tup: Union[Sequence[Array], Sequence[Quantity]]
-) -> Union[Array, Quantity]:
-  """
-  Stack 1-D arrays as columns into a 2-D array.
-
-  Take a sequence of 1-D arrays and stack them as columns to make a single
-  2-D array. 2-D arrays are stacked as-is, just like with hstack.
-
-  Parameters
-  ----------
-  tup : sequence of 1-D array_like, Quantity
-    1-D arrays to stack as columns.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    The array formed by stacking the given arrays.
-  """
-  return _fun_keep_unit_sequence(jnp.column_stack, tup)
-
-
-@set_module_as('brainunit.math')
-def block(
-    arrays: Sequence[Union[Array, Quantity]]
-) -> Union[Array, Quantity]:
-  """
-  Assemble a quantity or an array from nested lists of blocks.
-
-  Parameters
-  ----------
-  arrays : sequence of array_like, Quantity
-    Each element in `arrays` can itself be a nested sequence of arrays, in which case the blocks in the corresponding
-    cells are recursively stacked as the elements of the resulting array.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    The array constructed from the given blocks.
-  """
-  return _fun_keep_unit_sequence(jnp.block, arrays)
-
-
-@set_module_as('brainunit.math')
-def atleast_1d(
-    *arys: Union[Array, Quantity]
-) -> Union[Array, Quantity]:
-  """
-  View inputs as quantities or arrays with at least one dimension.
-
-  Parameters
-  ----------
-  *args : array_like, Quantity
-    One or more input arrays or quantities.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    An array or a quantity, or a tuple of arrays or quantities, each with `a.ndim >= 1`.
-  """
-  return _fun_keep_unit_sequence(jnp.atleast_1d, *arys)
-
-
-@set_module_as('brainunit.math')
-def atleast_2d(
-    *arys: Union[Array, Quantity]
-) -> Union[Array, Quantity]:
-  """
-  View inputs as quantities or arrays with at least two dimensions.
-
-  Parameters
-  ----------
-  *args : array_like, Quantity
-    One or more input arrays or quantities.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    An array or a quantity, or a tuple of arrays or quantities, each with `a.ndim >= 2`.
-  """
-  return _fun_keep_unit_sequence(jnp.atleast_2d, *arys)
-
-
-@set_module_as('brainunit.math')
-def atleast_3d(
-    *arys: Union[Array, Quantity]
-) -> Union[Array, Quantity]:
-  """
-  View inputs as quantities or arrays with at least three dimensions.
-
-  Parameters
-  ----------
-  *args : array_like, Quantity
-    One or more input arrays or quantities.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    An array or a quantity, or a tuple of arrays or quantities, each with `a.ndim >= 3`.
-  """
-  return _fun_keep_unit_sequence(jnp.atleast_3d, *arys)
-
-
-@set_module_as('brainunit.math')
-def append(
-    arr: Union[Array, Quantity],
-    values: Union[Array, Quantity],
-    axis: Optional[int] = None
-) -> Union[Array, Quantity]:
-  """
-  Append values to the end of a quantity or an array.
-
-  Parameters
-  ----------
-  arr : array_like, Quantity
-    Values are appended to a copy of this array.
-  values : array_like, Quantity
-    These values are appended to a copy of `arr`.
-    It must be of the correct shape (the same shape as `arr`, excluding `axis`).
-  axis : int, optional
-    The axis along which `values` are appended. If `axis` is None, `values` is flattened before use.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    A copy of `arr` with `values` appended to `axis`. Note that `append` does not occur in-place:
-    a new array is allocated and filled.
-  """
-  return _fun_keep_unit_sequence(jnp.append, arr, values, axis=axis)
-
-
-def _fun_keep_unit_return_sequence(
-    func,
-    x: jax.typing.ArrayLike | Quantity,
-    *args,
-    **kwargs
-):
-  if isinstance(x, Quantity):
-    r = func(x.value, *args, **kwargs)
-    return [Quantity(rr, dim=x.dim) for rr in r]
-  return func(x, *args, **kwargs)
-
-
-@set_module_as('brainunit.math')
-def split(
-    a: Union[Array, Quantity],
-    indices_or_sections: Union[int, Sequence[int]],
-    axis: int = 0
-) -> Union[List[Array], List[Quantity]]:
-  """
-  Split quantity or array into a list of multiple sub-arrays.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Array to be divided into sub-arrays.
-  indices_or_sections : int or 1-D array
-    If `indices_or_sections` is an integer, N, the array will be divided into
-    N equal arrays along `axis`. If such a split is not possible, an error is
-    raised. If `indices_or_sections` is a 1-D array of sorted integers, the
-    entries indicate where along `axis` the array is split. For example,
-    `[2, 3]` would, for `axis=0`, result in
-    - `a[:2]`
-    - `a[2:3]`
-    - `a[3:]`
-  axis : int, optional
-    The axis along which to split, default is 0.
-
-  Returns
-  -------
-  res : list of ndarrays, Quantity
-    A list of sub-arrays.
-  """
-  return _fun_keep_unit_return_sequence(jnp.split, a, indices_or_sections=indices_or_sections, axis=axis)
-
-
-@set_module_as('brainunit.math')
-def array_split(
-    ary: Union[Quantity, jax.typing.ArrayLike],
-    indices_or_sections: Union[int, jax.typing.ArrayLike],
-    axis: Optional[int] = 0
-) -> Union[list[Quantity], list[Array]]:
-  """
-  Split an array into multiple sub-arrays.
-
-  Parameters
-  ----------
-  ary : Quantity or array
-    Array to be divided into sub-arrays.
-  indices_or_sections : int or 1-D array
-    If `indices_or_sections` is an integer, `ary` is divided into `indices_or_sections` sub-arrays along `axis`.
-    If such a split is not possible, an error is raised.
-    If `indices_or_sections` is a 1-D array of sorted integers, the entries indicate where along `axis` the array is split.
-  axis : int, optional
-    The axis along which to split, default is 0.
-
-  Returns
-  -------
-  sub-arrays : list of Quantity or list of array
-    A list of sub-arrays.
-  """
-  return _fun_keep_unit_return_sequence(jnp.split, ary, indices_or_sections=indices_or_sections, axis=axis)
-
-
-@set_module_as('brainunit.math')
-def dsplit(
-    a: Union[Array, Quantity],
-    indices_or_sections: Union[int, Sequence[int]]
-) -> Union[List[Array], List[Quantity]]:
-  """
-  Split a quantity or an array into multiple sub-arrays along the 3rd axis (depth).
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Array to be divided into sub-arrays.
-  indices_or_sections : int or 1-D array
-    If `indices_or_sections` is an integer, N, the array will be divided into
-    N equal arrays along the third axis (depth). If such a split is not possible,
-    an error is raised. If `indices_or_sections` is a 1-D array of sorted integers,
-    the entries indicate where along the third axis the array is split.
-
-  Returns
-  -------
-  res : list of ndarrays, Quantity
-    A list of sub-arrays.
-  """
-  return _fun_keep_unit_return_sequence(jnp.dsplit, a, indices_or_sections)
-
-
-@set_module_as('brainunit.math')
-def hsplit(
-    a: Union[Array, Quantity],
-    indices_or_sections: Union[int, Sequence[int]]
-) -> Union[List[Array], List[Quantity]]:
-  """
-  Split a quantity or an array into multiple sub-arrays horizontally (column-wise).
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Array to be divided into sub-arrays.
-  indices_or_sections : int or 1-D array
-    If `indices_or_sections` is an integer, N, the array will be divided into
-    N equal arrays along the second axis. If such a split is not possible, an
-    error is raised. If `indices_or_sections` is a 1-D array of sorted integers,
-    the entries indicate where along the second axis the array is split.
-
-  Returns
-  -------
-  res : list of ndarrays, Quantity
-    A list of sub-arrays.
-  """
-  return _fun_keep_unit_return_sequence(jnp.hsplit, a, indices_or_sections)
-
-
-@set_module_as('brainunit.math')
-def vsplit(
-    a: Union[Array, Quantity],
-    indices_or_sections: Union[int, Sequence[int]]
-) -> Union[List[Array], List[Quantity]]:
-  """
-  Split a quantity or an array into multiple sub-arrays vertically (row-wise).
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Array to be divided into sub-arrays.
-  indices_or_sections : int or 1-D array
-    If `indices_or_sections` is an integer, N, the array will be divided into
-    N equal arrays along the first axis. If such a split is not possible, an
-    error is raised. If `indices_or_sections` is a 1-D array of sorted integers,
-    the entries indicate where along the first axis the array is split.
-
-  Returns
-  -------
-  res : list of ndarrays, Quantity
-    A list of sub-arrays.
-  """
-  return _fun_keep_unit_return_sequence(jnp.vsplit, a, indices_or_sections)
-
-
-@set_module_as('brainunit.math')
-def unique(
-    a: Union[Array, Quantity],
-    return_index: bool = False,
-    return_inverse: bool = False,
-    return_counts: bool = False,
-    axis: Optional[int] = None,
-    *,
-    equal_nan: bool = False,
-    size: Optional[int] = None,
-    fill_value: Optional[jax.typing.ArrayLike, Quantity] = None
-) -> Tuple[Array | Quantity] | Array | Quantity:
-  """
-  Find the unique elements of a quantity or an array.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Input array.
-  return_index : bool, optional
-    If True, also return the indices of `a` (along the specified axis, if provided) that result in the unique array.
-  return_inverse : bool, optional
-    If True, also return the indices of the unique array (for the specified axis, if provided)
-    that can be used to reconstruct `a`.
-  return_counts : bool, optional
-    If True, also return the number of times each unique item appears in `a`.
-  axis : int, optional
-    The axis along which to operate. If None, the array is flattened before use. Default is None.
-  equal_nan : bool, optional
-    Whether to compare NaN's as equal. If True, NaN's in `a` will be considered equal to each other in the unique array.
-  size : int, optional
-    The length of the output array. If `size` is not None, the output array will have the length of `size`.
-  fill_value : scalar, optional
-    The value to use for missing values. If `fill_value` is not None, the output array will have the length of `size`.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    The sorted unique values.
-  """
-  if isinstance(a, Quantity):
-    if fill_value is not None:
-      fail_for_dimension_mismatch(fill_value, a)
-      fill_value = fill_value.value
-    result = jnp.unique(a.value,
-                        return_index=return_index,
-                        return_inverse=return_inverse,
-                        return_counts=return_counts,
-                        axis=axis, equal_nan=equal_nan,
-                        size=size,
-                        fill_value=fill_value)
-    if isinstance(result, tuple):
-      output = []
-      output.append(Quantity(result[0], dim=a.dim))
-      for r in result[1:]:
-        output.append(r)
-      return tuple(output)
-    else:
-      return Quantity(result, dim=a.dim)
-  else:
-    if isinstance(fill_value, Quantity):
-      assert fill_value.is_unitless, 'fill_value must be unitless when "a" is not a Quantity.'
-      fill_value = fill_value.value
-    return jnp.unique(a,
-                      return_index=return_index,
-                      return_inverse=return_inverse,
-                      return_counts=return_counts,
-                      axis=axis,
-                      equal_nan=equal_nan,
-                      size=size,
-                      fill_value=fill_value)
-
-
-@set_module_as('brainunit.math')
-def compress(
-    condition: Array,
-    a: Union[Array, Quantity],
-    axis: Optional[int] = None,
-    *,
-    size: Optional[int] = None,
-    fill_value: Optional[jax.typing.ArrayLike] = None,
-) -> Union[Array, Quantity]:
-  """
-  Return selected slices of a quantity or an array along given axis.
-
-  Parameters
-  ----------
-  condition : array_like, Quantity
-    An array of boolean values that selects which slices to return. If the shape of condition is not the same as `a`,
-    it must be broadcastable to `a`.
-  a : array_like, Quantity
-    Array from which to extract a part.
-  axis : int or None, optional
-    The axis along which to take slices. If axis is None, `condition` must be a 1-D array with the same length as `a`.
-    If axis is an integer, `condition` must be broadcastable to the same shape as `a` along all axes except `axis`.
-  size : int, optional
-    The length of the returned axis. By default, the length of the input array along the axis is used.
-  fill_value : scalar, optional
-    The value to use for elements in the output array that are not selected. If None, the output array has the same
-    type as `a` and is filled with zeros.
-
-  Returns
-  -------
-  res : ndarray, Quantity
-    A new array that has the same number of dimensions as `a`, and the same shape as `a` with axis `axis` removed.
-  """
-  assert not isinstance(condition, Quantity), f'condition must be an array_like. But got {condition}'
-  if isinstance(a, Quantity):
-    if fill_value is not None:
-      fail_for_dimension_mismatch(fill_value, a)
-      fill_value = fill_value.value
-  else:
-    if isinstance(fill_value, Quantity):
-      assert fill_value.is_unitless, 'fill_value must be unitless when "a" is not a Quantity.'
-      fill_value = fill_value.value
-  return _fun_keep_unit_unary(functools.partial(jnp.compress, condition),
-                              a, axis=axis, size=size, fill_value=fill_value)
-
-
-@set_module_as('brainunit.math')
-def extract(
-    condition: Array,
-    arr: Union[Array, Quantity],
-    *,
-    size: Optional[int] = None,
-    fill_value: Optional[jax.typing.ArrayLike | Quantity] = None,
-) -> Array | Quantity:
-  """
-  Return the elements of an array that satisfy some condition.
-
-  Parameters
-  ----------
-  condition : array_like, Quantity
-    An array of boolean values that selects which elements to extract.
-  arr : array_like, Quantity
-    The array from which to extract elements.
-  size: int
-    optional static size for output. Must be specified in order for ``extract``
-    to be compatible with JAX transformations like :func:`~jax.jit` or :func:`~jax.vmap`.
-  fill_value: array_like
-    if ``size`` is specified, fill padded entries with this value (default: 0).
-
-  Returns
-  -------
-  res : ndarray
-    The extracted elements. The shape of `res` is the same as that of `condition`.
-  """
-  assert not isinstance(condition, Quantity), f'condition must be an array_like. But got {condition}'
-  if isinstance(arr, Quantity):
-    if fill_value is not None:
-      fail_for_dimension_mismatch(fill_value, arr)
-      fill_value = fill_value.value
-  else:
-    if isinstance(fill_value, Quantity):
-      assert fill_value.is_unitless, 'fill_value must be unitless when "a" is not a Quantity.'
-      fill_value = fill_value.value
-  return _fun_keep_unit_unary(functools.partial(jnp.extract, condition),
-                              arr, size=size, fill_value=fill_value)
-
-
-@set_module_as('brainunit.math')
-def argsort(
-    a: Union[Array, Quantity],
-    axis: Optional[int] = -1,
-    *,
-    kind: None = None,
-    order: None = None,
-    stable: bool = True,
-    descending: bool = False,
-) -> Array:
-  """
-  Returns the indices that would sort an array or a quantity.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Array or quantity to be sorted.
-  axis : int or None, optional
-    Axis along which to sort. If None, the array is flattened before sorting. The default is -1, which sorts along
-    the last axis.
-  kind : {'quicksort', 'mergesort', 'heapsort', 'stable'}, optional
-    Sorting algorithm. The default is 'None'.
-  order : str or list of str, optional
-    When `a` is a quantity, it can be a string or a sequence of strings, which is interpreted as an order the quantity
-    should be sorted. The default is None.
-  stable : bool, optional
-    Whether to use a stable sorting algorithm. The default is True.
-  descending : bool, optional
-    Whether to sort in descending order. The default is False.
-
-  Returns
-  -------
-  res : ndarray
-    Array of indices that sort the array.
-  """
-  return _fun_remove_unit_unary(jnp.argsort,
-                                a,
-                                axis=axis,
-                                kind=kind,
-                                order=order,
-                                stable=stable,
-                                descending=descending)
-
-
-@set_module_as('brainunit.math')
-def argmax(
-    a: Union[Array, Quantity],
-    axis: Optional[int] = None,
-    keepdims: Optional[bool] = None
-) -> Array:
-  """
-  Returns indices of the max value along an axis.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Input data.
-  axis : int, optional
-    By default, the index is into the flattened array, otherwise along the specified axis.
-  keepdims : bool, optional
-    If this is set to True, the axes which are reduced are left in the result as dimensions with size one. With this
-    option, the result will broadcast correctly against the input array.
-
-  Returns
-  -------
-  res : ndarray
-    Array of indices into the array. It has the same shape as `a.shape` with the dimension along `axis` removed.
-  """
-  return _fun_remove_unit_unary(jnp.argmax, a, axis=axis, keepdim=keepdims)
-
-
-@set_module_as('brainunit.math')
-def argmin(
-    a: Union[Array, Quantity],
-    axis: Optional[int] = None,
-    keepdims: Optional[bool] = None
-) -> Array:
-  """
-  Returns indices of the min value along an axis.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Input data.
-  axis : int, optional
-    By default, the index is into the flattened array, otherwise along the specified axis.
-  keepdims : bool, optional
-    If this is set to True, the axes which are reduced are left in the result as dimensions with size one. With this
-    option, the result will broadcast correctly against the input array.
-
-  Returns
-  -------
-  res : ndarray
-    Array of indices into the array. It has the same shape as `a.shape` with the dimension along `axis` removed.
-  """
-  return _fun_remove_unit_unary(jnp.argmin, a, axis=axis, keepdims=keepdims)
-
-
-@set_module_as('brainunit.math')
-def nanargmax(
-    a: Union[jax.typing.ArrayLike, Quantity],
-    axis: int = None,
-    keepdims: bool = False
-) -> jax.Array:
-  """
-  Return the indices of the maximum values in the specified axis ignoring
-  NaNs. For all-NaN slices ``ValueError`` is raised. Warning: the
-  results cannot be trusted if a slice contains only NaNs and -Infs.
-
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Input data.
-  axis : int, optional
-    Axis along which to operate.  By default flattened input is used.
-  keepdims : bool, optional
-    If this is set to True, the axes which are reduced are left
-    in the result as dimensions with size one. With this option,
-    the result will broadcast correctly against the array.
-
-  Returns
-  -------
-  index_array : ndarray
-    An array of indices or a single index value.
-  """
-  return _fun_remove_unit_unary(jnp.nanargmax,
-                                a,
-                                return_quantity=False,
-                                axis=axis,
-                                keepdims=keepdims)
-
-
-@set_module_as('brainunit.math')
-def nanargmin(
-    a: Union[jax.typing.ArrayLike, Quantity],
-    axis: int = None,
-    keepdims: bool = False
-) -> jax.Array:
-  """
-  Return the indices of the minimum values in the specified axis ignoring
-  NaNs. For all-NaN slices ``ValueError`` is raised. Warning: the results
-  cannot be trusted if a slice contains only NaNs and Infs.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Input data.
-  axis : int, optional
-    Axis along which to operate.  By default flattened input is used.
-  keepdims : bool, optional
-    If this is set to True, the axes which are reduced are left
-    in the result as dimensions with size one. With this option,
-    the result will broadcast correctly against the array.
-
-  Returns
-  -------
-  index_array : ndarray
-    An array of indices or a single index value.
-  """
-  return _fun_remove_unit_unary(jnp.nanargmin,
-                                a,
-                                return_quantity=False,
-                                axis=axis,
-                                keepdims=keepdims)
-
-
-@set_module_as('brainunit.math')
-def argwhere(
-    a: Union[Array, Quantity],
-    *,
-    size: Optional[int] = None,
-    fill_value: Optional[jax.typing.ArrayLike] = None,
-) -> Array:
-  """
-  Find the indices of array elements that are non-zero, grouped by element.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Input data.
-  size : int, optional
-    The length of the returned axis. By default, the length of the input array along the axis is used.
-  fill_value : scalar, optional
-    The value to use for elements in the output array that are not selected. If None, the output array has the same
-    type as `a` and is filled with zeros.
-
-  Returns
-  -------
-  res : ndarray
-    The indices of elements that are non-zero. The indices are grouped by element.
-  """
-  return _fun_remove_unit_unary(jnp.argwhere, a, size=size, fill_value=fill_value)
-
-
-@set_module_as('brainunit.math')
-def nonzero(
-    a: Union[Array, Quantity],
-    *,
-    size: Optional[int] = None,
-    fill_value: Optional[jax.typing.ArrayLike] = None,
-) -> Tuple[Array, ...]:
-  """
-  Return the indices of the elements that are non-zero.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Input data.
-  size : int, optional
-    The length of the returned axis. By default, the length of the input array along the axis is used.
-  fill_value : scalar, optional
-    The value to use for elements in the output array that are not selected. If None, the output array has the same
-    type as `a` and is filled with zeros.
-
-  Returns
-  -------
-  res : tuple of ndarrays
-    Indices of elements that are non-zero along the specified axis. Each array in the tuple has the same shape as the
-    input array.
-  """
-  return _fun_remove_unit_unary(jnp.nonzero, a, size=size, fill_value=fill_value)
-
-
-@set_module_as('brainunit.math')
-def flatnonzero(
-    a: Union[Array, Quantity],
-    *,
-    size: Optional[int] = None,
-    fill_value: Optional[jax.typing.ArrayLike] = None,
-) -> Array:
-  """
-  Return indices that are non-zero in the flattened version of the input quantity or array.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Input data.
-  size : int, optional
-    The length of the returned axis. By default, the length of the input array along the axis is used.
-  fill_value : scalar, optional
-    The value to use for elements in the output array that are not selected. If None, the output array has the same
-    type as `a` and is filled with zeros.
-
-  Returns
-  -------
-  res : ndarray
-    Output array, containing the indices of the elements of `a.ravel()` that are non-zero.
-  """
-  return _fun_remove_unit_unary(jnp.flatnonzero, a, size=size, fill_value=fill_value)
-
-
-@set_module_as('brainunit.math')
-def count_nonzero(
-    a: Union[Array, Quantity],
-    axis: Optional[int] = None,
-    keepdims: Optional[bool] = None
-) -> Array:
-  """
-  Count the number of non-zero values in the quantity or array `a`.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    The array for which to count non-zeros.
-  axis : int, optional
-    The axis along which to count the non-zeros. If `None`, count non-zeros over the entire array.
-  keepdims : bool, optional
-    If this is set to `True`, the axes which are counted are left in the result as dimensions with size one. With this
-    option, the result will broadcast correctly against the original array.
-
-  Returns
-  -------
-  res : ndarray
-    Number of non-zero values in the quantity or array along a given axis.
-  """
-  return _fun_remove_unit_unary(jnp.count_nonzero, a, axis=axis, keepdims=keepdims)
-
-
-@set_module_as('brainunit.math')
-def searchsorted(
-    a: Union[Array, Quantity],
-    v: Union[Array, Quantity],
-    side: str = 'left',
-    sorter: Optional[Array] = None,
-    *,
-    method: Optional[str] = 'scan'
-) -> Array:
-  """
-  Find indices where elements should be inserted to maintain order.
-
-  Find the indices into a sorted array `a` such that, if the corresponding elements in `v` were inserted before the
-  indices, the order of `a` would be preserved.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Input array. It must be sorted in ascending order.
-  v : array_like, Quantity
-    Values to insert into `a`.
-  side : {'left', 'right'}, optional
-    If 'left', the index of the first suitable location found is given. If 'right', return the last such index. If
-    there is no suitable index, return either 0 or N (where N is the length of `a`).
-  sorter : 1-D array_like, optional
-    Optional array of integer indices that sort array `a` into ascending order. They are typically the result of
-    `argsort`.
-  method : str
-    One of 'scan' (default), 'scan_unrolled', 'sort' or 'compare_all'. Controls the method used by the
-    implementation: 'scan' tends to be more performant on CPU (particularly when ``a`` is
-    very large), 'scan_unrolled' is more performant on GPU at the expense of additional compile time,
-    'sort' is often more performant on accelerator backends like GPU and TPU
-    (particularly when ``v`` is very large), and 'compare_all' can be most performant
-    when ``a`` is very small. The default is 'scan'.
-
-  Returns
-  -------
-  out : ndarray
-    Array of insertion points with the same shape as `v`.
-  """
-  if isinstance(a, Quantity):
-    fail_for_dimension_mismatch(a, v)
-    a = a.value
-    v = v.value
-  if isinstance(v, Quantity):
-    assert v.is_unitless, 'v must be unitless when "a" is not a Quantity.'
-    v = v.value
-  return jnp.searchsorted(a, v, side=side, sorter=sorter, method=method)
